#23 move all React components to the component folder
